backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API for musikkhylla music rack application"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "pydantic>=2.10.0",
    "email-validator>=2.1.0",
]

[project.scripts]
dev = "app:main"

[tool.uv]
dev-dependencies = []
